chore: restructure spider & add login tips
