--- core/spider/src/websites/x/grabs/index.ts
@@ -1,1 +1,0 @@
-export * as UserPage from './user'
