+++ core/spider/src/websites/x/index.ts
@@ -1,1 +1,1 @@
-export * as TweetGrabber from './grabs'
+export * as UserPage from './user'
